refactor: Split SLiceHeaderControls into smaller files
